added ts-node-dev and Mutation
