+++ package.json
@@ -10,10 +10,12 @@
    "devDependencies": {
       "@types/node": "^16.11.10",
       "ts-node": "10.7.0",
+      "ts-node-dev": "^2.0.0",
       "typescript": "4.5.2"
    },
    "scripts": {
       "start": "ts-node src/index.ts",
+      "start:dev":"ts-node-dev --respawn --transpile-only src/index.ts",
       "typeorm": "typeorm-ts-node-commonjs"
    }
 }
