+++ src/index.ts
@@ -1,23 +1,47 @@
 import "reflect-metadata";
 import { AppDataSource } from "./data-source";
-import { User } from "./entity/User";
 
-AppDataSource.initialize()
-  .then(async () => {
-    console.log("Inserting a new user into the database...");
-    const user = new User();
-    user.firstName = "Samuel";
-    user.lastName = "Santana";
-    user.age = 20;
-    await AppDataSource.manager.save(user);
-    console.log("Saved a new user with id: " + user.id);
+import { ApolloServer, gql } from "apollo-server";
 
-    console.log("Loading users from the database...");
-    const users = await AppDataSource.manager.find(User);
-    console.log("Loaded users: ", users);
+const typeDefs = gql`
+  input UserInput {
+    name: String!
+    email: String!
+    password: String!
+    birthDate: String!
+  }
 
-    console.log(
-      "Here you can setup and run express / fastify / any other framework."
-    );
-  })
-  .catch((error) => console.log(error));
+  type User {
+    id: Int
+    name: String!
+    email: String!
+  }
+
+  type Query {
+    hello: String
+  }
+
+  type Mutation {
+    createUser(data: UserInput!): String
+  }
+`;
+
+const resolvers = {
+  Query: {
+    hello: () => "Hello Word!",
+  },
+  Mutation: {
+    createUser: (parent, args) => {
+      const data = { id: 1, name: args.data.name, email: args.data.email };
+      return `Usuario criado ${data.name}`;
+    },
+  },
+};
+
+const port = 4000;
+AppDataSource.initialize();
+const app = new ApolloServer({ typeDefs, resolvers });
+app
+  .listen(port)
+  .then(() => console.info(`Server executing on port ${port}`))
+  .catch((error) => console.error(error));
